Update used  method for panel, update tests
