+++ public/app/plugins/panel/logs/LogsPanel.tsx
@@ -3,7 +3,6 @@ import { LogRows, CustomScrollbar } from '@grafana/ui';
 import { LogsDedupStrategy, PanelProps } from '@grafana/data';
 import { Options } from './types';
 import { dataFrameToLogsModel } from 'app/core/logs_model';
-import { sortLogsResult } from 'app/core/utils/explore';
 
 interface LogsPanelProps extends PanelProps<Options> {}
 
@@ -22,12 +21,11 @@ export const LogsPanel: React.FunctionComponent<LogsPanelProps> = ({
   }
 
   const newResults = data ? dataFrameToLogsModel(data.series, data.request?.intervalMs, timeZone) : null;
-  const sortedNewResults = sortLogsResult(newResults, sortOrder);
 
   return (
     <CustomScrollbar autoHide>
       <LogRows
-        logRows={sortedNewResults.rows}
+        logRows={newResults?.rows || []}
         dedupStrategy={LogsDedupStrategy.none}
         highlighterExpressions={[]}
         showLabels={showLabels}
@@ -36,6 +34,7 @@ export const LogsPanel: React.FunctionComponent<LogsPanelProps> = ({
         timeZone={timeZone}
         allowDetails={true}
         disableCustomHorizontalScroll={true}
+        logsOrder={sortOrder}
       />
     </CustomScrollbar>
   );
